+++ src/main/java/com/example/useful/repository/EmpBankDetailsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.useful.entity.EmpBankDetails;
 
-public interface EmpBankDetailsRepository extends JpaRepository<EmpBankDetails,Integer>{
+public interface EmpBankDetailsRepository extends JpaRepository<EmpBankDetails,Long>{
 
 	EmpBankDetails findByAccountNo(String accountNumber);
